moved handler for the mods part into its own file
